requirements.txt
boto3==1.35.47
mysql-connector-python==9.0.0
Flask==3.0.3